Cleanly relinquish template file resources
Currently, on rendering a template that is read from a file,
the file resources are not cleanly relinquished. This change
adds an additional attribute to `lwan_tpl_t` type that repre-
sents that file descriptor of the file opened. When freeing up
`lwan_tpl_t` after usage, this decriptor if present, is closed.
If the template is read from a string, this attribute is ignored.